+++ r_main.cpp
@@ -295,7 +295,7 @@ R_PointToAngle
 	    if (x>y)
 	    {
 		// octant 8
-		return -tantoangle[SlopeDiv(y,x)];
+		return -static_cast<int>(tantoangle[SlopeDiv(y,x)]);
 	    }
 	    else
 	    {
