+++ r_segs.cpp
@@ -590,7 +590,7 @@ R_StoreWallRange
 	offsetangle = rw_normalangle-rw_angle1;
 	
 	if (offsetangle > ANG180)
-	    offsetangle = -offsetangle;
+	    offsetangle = -static_cast<int>(offsetangle);
 
 	if (offsetangle > ANG90)
 	    offsetangle = ANG90;
